\name{function-internal}
\alias{error.est}
\alias{get.subset}
\alias{hard.thresh}
\alias{hard.thresh.scaler}
\alias{holm.robust}
\alias{huberize}
\alias{soft.thresh}
\alias{soft.thresh.scaler}
\alias{subsp.dist.orth}
\title{
Internal ssvd functions.
}
\description{
Internal functions called by ssvd, ssvd.initial, and ssvd.iter.thresh.
}
\details{
These are not to be called directly by the users.
}
\references{
A Sparse SVD Method for High-dimensional Data
}
\author{
Dan Yang
}
\keyword{internal}